fixed memory leaks in evaluate and clickonspecificbutton
the button problem was that the pause button was not initialized
and clickonspecificbutton tried to reach it.
solution: set pause button to NULL in the buttons array when the game
is over, and the function that goes over the buttons ignores
buttons that are NULL.

evaluate: didn't free the dist matrix memory when the game is over.
solution - create dist matrix only if it is not a game over state.